# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task Manager API",
    "version": "1.1.0",
    "description": "Advanced REST API for authentication, user management, task management with comments, attachments, subtasks, recurring tasks, activity feed, Kanban boards, notifications, and team collaboration features."
  },
  "servers": [
    { "url": "/api" }
  ],
  "components": {
    "securitySchemes": {
      "ApiToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-auth-token"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "Validation error" },
          "errors": {
            "type": "array",
            "items": { "type": "object" }
          }
        }
      },
...
      "PerformanceMetrics": {
        "type": "object",
        "properties": {
          "requests": {
            "type": "object",
            "properties": {
              "total": { "type": "number", "example": 1250 },
              "byMethod": {
                "type": "object",
                "properties": {
                  "GET": { "type": "number", "example": 800 },
                  "POST": { "type": "number", "example": 300 },
                  "PATCH": { "type": "number", "example": 100 },
                  "DELETE": { "type": "number", "example": 50 }
                }
              },
              "byRoute": {
                "type": "object",
                "additionalProperties": { "type": "number" }
              },
              "byStatusCode": {
                "type": "object",
                "properties": {
                  "200": { "type": "number", "example": 1100 },
                  "201": { "type": "number", "example": 80 },
                  "400": { "type": "number", "example": 30 },
                  "401": { "type": "number", "example": 25 },
                  "404": { "type": "number", "example": 10 },
                  "500": { "type": "number", "example": 5 }
                }
              }
            }
          },
          "responseTimes": {
            "type": "object",
            "properties": {
              "avgResponseTime": { "type": "number", "example": 125.5 },
              "p95ResponseTime": { "type": "number", "example": 280.2 },
              "p99ResponseTime": { "type": "number", "example": 450.8 },
              "min": { "type": "number", "example": 12.3 },
              "max": { "type": "number", "example": 1200.5 }
            }
          },
          "memory": {
            "type": "object",
            "properties": {
              "heapUsed": { "type": "number", "example": 45678901 },
              "heapTotal": { "type": "number", "example": 67890123 },
              "external": { "type": "number", "example": 1234567 },
              "rss": { "type": "number", "example": 89012345 }
            }
          },
          "calculated": {
            "type": "object",
            "properties": {
              "requestsPerSecond": { "type": "number", "example": 15.8 },
              "errorRate": { "type": "number", "example": 2.5 }
            }
          },
          "uptime": { "type": "number", "example": 86400 },
          "timestamp": { "type": "string", "format": "date-time", "example": "2025-01-15T10:30:00.000Z" }
        }
      },
...
      "BenchmarkRecommendations": {
        "type": "object",
        "properties": {
          "currentMetrics": { "$ref": "#/components/schemas/PerformanceMetrics" },
          "recommendations": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "category": {
                  "type": "string",
                  "enum": ["database", "memory", "response_time", "error_rate", "caching"],
                  "example": "response_time"
                },
                "priority": {
                  "type": "string",
                  "enum": ["low", "medium", "high", "critical"],
                  "example": "high"
                },
                "title": { "type": "string", "example": "Optimize slow database queries" },
                "description": { "type": "string", "example": "Average response time is above recommended threshold. Consider adding database indexes or implementing caching." },
                "action": { "type": "string", "example": "Add indexes to frequently queried fields" }
              }
            }
          },
          "benchmarks": {
            "type": "object",
            "properties": {
              "responseTimes": {
                "type": "object",
                "properties": {
                  "excellent": { "type": "number", "example": 100 },
                  "good": { "type": "number", "example": 300 },
                  "acceptable": { "type": "number", "example": 1000 },
                  "poor": { "type": "number", "example": 1001 }
                }
              },
              "errorRates": {
                "type": "object",
                "properties": {
                  "excellent": { "type": "number", "example": 0.1 },
                  "good": { "type": "number", "example": 1.0 },
                  "acceptable": { "type": "number", "example": 5.0 },
                  "poor": { "type": "number", "example": 5.1 }
                }
              },
              "throughput": {
                "type": "object",
                "properties": {
                  "minimum": { "type": "number", "example": 10 },
                  "target": { "type": "number", "example": 50 },
                  "excellent": { "type": "number", "example": 100 }
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    { "ApiToken": [] },
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/performance/reset": {
      "post": {
        "summary": "Reset performance metrics (admin only)",
        "tags": ["Performance"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Performance metrics reset successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Performance metrics reset successfully" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/performance/benchmark": {
      "get": {
        "summary": "Get benchmark recommendations",
        "tags": ["Performance"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Benchmark recommendations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/BenchmarkRecommendations" },
                    "message": { "type": "string", "example": "Performance benchmark recommendations retrieved successfully" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  }
}